Extract attribute types to separate module
